// types/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Dimension: any;
  HexColor: any;
  JSON: any;
  Quality: any;
};

/** [See type definition](https://app.contentful.com/spaces/vvmytyavfefi/content_types/article) */
export type Article = Entry & {
  __typename?: 'Article';
  content?: Maybe<ArticleContent>;
  contentfulMetadata: ContentfulMetadata;
  countriesCollection?: Maybe<ArticleCountriesCollection>;
  coverImage?: Maybe<Asset>;
  description?: Maybe<Scalars['String']>;
  industriesCollection?: Maybe<ArticleIndustriesCollection>;
  linkedFrom?: Maybe<ArticleLinkingCollections>;
  relatedArticlesCollection?: Maybe<ArticleRelatedArticlesCollection>;
  slug?: Maybe<Scalars['String']>;
  sys: Sys;
  title?: Maybe<Scalars['String']>;
};
// ...
export type ArticleCollection = {
  __typename?: 'ArticleCollection';
  items: Array<Maybe<Article>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};

export type ArticleContent = {
  __typename?: 'ArticleContent';
  json: Scalars['JSON'];
  links: ArticleContentLinks;
};

export type ArticleContentAssets = {
  __typename?: 'ArticleContentAssets';
  block: Array<Maybe<Asset>>;
  hyperlink: Array<Maybe<Asset>>;
};

export type ArticleContentEntries = {
  __typename?: 'ArticleContentEntries';
  block: Array<Maybe<Entry>>;
  hyperlink: Array<Maybe<Entry>>;
  inline: Array<Maybe<Entry>>;
};

export type ArticleContentLinks = {
  __typename?: 'ArticleContentLinks';
  assets: ArticleContentAssets;
  entries: ArticleContentEntries;
};

export type ArticleCountriesCollection = {
  __typename?: 'ArticleCountriesCollection';
  items: Array<Maybe<Country>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type ArticleIndustriesCollection = {
  __typename?: 'ArticleIndustriesCollection';
  items: Array<Maybe<Industry>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};

export type ArticleLinkingCollections = {
  __typename?: 'ArticleLinkingCollections';
  articleCollection?: Maybe<ArticleCollection>;
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type ArticleRelatedArticlesCollection = {
  __typename?: 'ArticleRelatedArticlesCollection';
  items: Array<Maybe<Article>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};

/** Represents a binary file in a space. An asset can be any file type. */
export type Asset = {
  __typename?: 'Asset';
  contentType?: Maybe<Scalars['String']>;
  contentfulMetadata: ContentfulMetadata;
  description?: Maybe<Scalars['String']>;
  fileName?: Maybe<Scalars['String']>;
  height?: Maybe<Scalars['Int']>;
  linkedFrom?: Maybe<AssetLinkingCollections>;
  size?: Maybe<Scalars['Int']>;
  sys: Sys;
  title?: Maybe<Scalars['String']>;
  url?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
};
// ...
export type AssetLinkingCollections = {
  __typename?: 'AssetLinkingCollections';
  articleCollection?: Maybe<ArticleCollection>;
  entryCollection?: Maybe<EntryCollection>;
};
// ...
export type ContentfulMetadata = {
  __typename?: 'ContentfulMetadata';
  tags: Array<Maybe<ContentfulTag>>;
};
// ...
export type ContentfulTag = {
  __typename?: 'ContentfulTag';
  id?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
};

/** [See type definition](https://app.contentful.com/spaces/vvmytyavfefi/content_types/country) */
export type Country = Entry & {
  __typename?: 'Country';
  code?: Maybe<Scalars['String']>;
  contentfulMetadata: ContentfulMetadata;
  linkedFrom?: Maybe<CountryLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  sys: Sys;
};
// ...
export type Entry = {
  contentfulMetadata: ContentfulMetadata;
  sys: Sys;
};

export type EntryCollection = {
  __typename?: 'EntryCollection';
  items: Array<Maybe<Entry>>;
  limit: Scalars['Int'];
  skip: Scalars['Int'];
  total: Scalars['Int'];
};
// ...
export type Industry = Entry & {
  __typename?: 'Industry';
  contentfulMetadata: ContentfulMetadata;
  linkedFrom?: Maybe<IndustryLinkingCollections>;
  name?: Maybe<Scalars['String']>;
  sys: Sys;
};
// ...
export type Sys = {
  __typename?: 'Sys';
  environmentId: Scalars['String'];
  firstPublishedAt?: Maybe<Scalars['DateTime']>;
  id: Scalars['String'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  publishedVersion?: Maybe<Scalars['Int']>;
  spaceId: Scalars['String'];
};